Redid the parser
This version is repetitive, but feels more organized to me. I have structs and parser functions for each lookup within a FAKEula query, and a function to slice the raw query, format it, and place all the data into a multi-leveled hash map with the raw data as the key value.